# apps/api/openapi.yaml
openapi: 3.1.0
info:
  title: QRForge API
  version: 0.1.0
  summary: Self-hosted API for dynamic QR codes, campaigns and privacy-friendly analytics.
  description: |
    The QRForge REST API. All endpoints live under `/api/v1` and exchange JSON with
    snake_case fields. Timestamps are RFC 3339 strings in UTC.

    ## Authentication
    - **API keys** (recommended for scripts): `Authorization: Bearer qrf_...`. Create keys in the
      dashboard (Settings → API keys). The full key is shown exactly once; only a hash is stored.
      API keys can manage QR codes, campaigns and analytics, and read domains. They cannot manage
      API keys, change account details or use admin endpoints.
    - **Browser sessions**: the `qrforge_session` HttpOnly cookie set by `/auth/login`. Every
      state-changing request made with a session must send the `X-CSRF-Token` header returned by
      `/auth/login`, `/auth/register` and `/auth/me`.

    ## Errors
    Errors use the HTTP status code plus a body of the form
    `{"error": {"code": "validation_failed", "message": "...", "fields": {"destination": "..."}}}`.

    ## Pagination
    List endpoints accept `page` (1-based) and `per_page` (1–100, default 20) and respond with
    `{"data": [...], "pagination": {"page", "per_page", "total"}}`.

    ## Rate limits
    Requests are rate limited per client IP and per API key. Exceeding a limit returns `429`
    with a `Retry-After` header.
  license:
    name: MIT
    identifier: MIT
servers:
...
security:
  - apiKey: []
  - session: []
paths:
...
  /auth/register:
    post:
      tags: [Auth]
      summary: Create an account
      description: The first account on an instance becomes an administrator and is always allowed, even when registration is closed.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string, minLength: 8, maxLength: 72, description: "8–72 bytes (bcrypt limit)" }
                name: { type: string, maxLength: 100 }
      responses:
        "201": { $ref: "#/components/responses/SessionResponse" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/login:
    post:
      tags: [Auth]
      summary: Log in
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email, password]
              properties:
                email: { type: string, format: email }
                password: { type: string }
      responses:
        "200": { $ref: "#/components/responses/SessionResponse" }
        "401": { $ref: "#/components/responses/Unauthorized" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/logout:
    post:
      tags: [Auth]
      summary: Log out (deletes the session)
      security: [{ session: [] }, {}]
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      responses:
        "204": { description: Logged out }
        "403": { $ref: "#/components/responses/Forbidden" }
  /auth/me:
    get:
      tags: [Auth]
      summary: Current user
      responses:
        "200":
          description: The authenticated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
                  csrf_token: { type: [string, "null"], description: "null when authenticated with an API key" }
        "401": { $ref: "#/components/responses/Unauthorized" }
    patch:
      tags: [Auth]
      summary: Update profile
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              properties:
                name: { type: string, maxLength: 100 }
                email: { type: string, format: email }
                current_password: { type: string, description: Required when changing the email address }
      responses:
        "200":
          description: Updated user
          content:
            application/json:
              schema:
                type: object
                properties:
                  user: { $ref: "#/components/schemas/User" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /auth/password/change:
    post:
      tags: [Auth]
      summary: Change password
      description: Revokes all other sessions.
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [current_password, new_password]
              properties:
                current_password: { type: string }
                new_password: { type: string, minLength: 8, maxLength: 72, description: "8–72 bytes (bcrypt limit)" }
      responses:
        "204": { description: Password changed }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /auth/password/forgot:
    post:
      tags: [Auth]
      summary: Request a password reset link
      description: |
        Always returns 202 so the response never reveals whether an account exists. When SMTP is
        configured the link is emailed; otherwise it is written to the server log for the
        administrator to deliver.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [email]
              properties:
                email: { type: string, format: email }
      responses:
        "202": { $ref: "#/components/responses/Accepted" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /auth/password/reset:
    post:
      tags: [Auth]
      summary: Reset a password with a token
      description: Tokens are single-use and expire after one hour. All sessions are revoked.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [token, password]
              properties:
                token: { type: string }
                password: { type: string, minLength: 8, maxLength: 72, description: "8–72 bytes (bcrypt limit)" }
      responses:
        "204": { description: Password reset }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /qrcodes:
    post:
      tags: [QR codes]
      summary: Create a QR code
      description: |
        Dynamic codes (default) encode a short redirect URL (`/r/{slug}`) whose destination can be
        changed later. Static codes store the encoded payload for re-downloading and organisation.
        `type` is accepted as an alias of `kind`.
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QRCodeCreate" }
            examples:
              dynamic:
                value: { name: Website, type: dynamic, destination: "https://example.com" }
              static:
                value: { name: Guest Wi-Fi, kind: static, qr_type: wifi, content: "WIFI:T:WPA;S:Cafe;P:espresso;;" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QRCode" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    get:
      tags: [QR codes]
      summary: List QR codes
      parameters:
        - $ref: "#/components/parameters/Page"
        - $ref: "#/components/parameters/PerPage"
        - { name: kind, in: query, schema: { type: string, enum: [static, dynamic] } }
        - { name: status, in: query, schema: { type: string, enum: [active, disabled] } }
        - { name: campaign_id, in: query, schema: { type: string } }
        - { name: q, in: query, description: Search name, destination, slug and content, schema: { type: string } }
        - name: sort
          in: query
          schema: { type: string, enum: [created_at, -created_at, name, -name, scan_count, -scan_count], default: -created_at }
      responses:
        "200":
          description: A page of QR codes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Paginated"
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: "#/components/schemas/QRCode" } }
  /qrcodes/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [QR codes]
      summary: Get a QR code
      responses:
        "200":
          description: The QR code
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QRCode" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [QR codes]
      summary: Update a QR code
      description: Only the provided fields change. Destination changes are recorded in the redirect history. Send `null` to clear nullable fields.
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/QRCodeUpdate" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QRCode" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [QR codes]
      summary: Delete a QR code (and its analytics)
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /qrcodes/{id}/rotate-slug:
    post:
      tags: [QR codes]
      summary: Replace the slug of a dynamic code
      description: The old slug stops resolving immediately. Printed codes using the old slug will no longer work.
      parameters: [{ $ref: "#/components/parameters/Id" }, { $ref: "#/components/parameters/CSRF" }]
      responses:
        "200":
          description: Updated code with a new slug
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QRCode" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /qrcodes/{id}/duplicate:
    post:
      tags: [QR codes]
      summary: Duplicate a QR code
      parameters: [{ $ref: "#/components/parameters/Id" }, { $ref: "#/components/parameters/CSRF" }]
      responses:
        "201":
          description: The copy (new slug, no scans)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/QRCode" }
        "403": { $ref: "#/components/responses/Forbidden" }
        "404": { $ref: "#/components/responses/NotFound" }
  /qrcodes/{id}/history:
    get:
      tags: [QR codes]
      summary: Destination history (newest first)
      parameters: [{ $ref: "#/components/parameters/Id" }]
      responses:
        "200":
          description: History entries
          content:
            application/json:
              schema:
                type: object
                properties:
                  data:
                    type: array
                    items:
                      type: object
                      properties:
                        destination: { type: string }
                        previous_destination: { type: [string, "null"] }
                        changed_at: { type: string, format: date-time }
                        changed_by_email: { type: [string, "null"] }
        "404": { $ref: "#/components/responses/NotFound" }
  /qrcodes/{id}/analytics:
    get:
      tags: [Analytics]
      summary: Analytics for one QR code
      parameters: [{ $ref: "#/components/parameters/Id" }, { $ref: "#/components/parameters/Range" }]
      responses:
        "200":
          description: Analytics report
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Analytics" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /analytics/overview:
    get:
      tags: [Analytics]
      summary: Analytics across all of your QR codes
      parameters: [{ $ref: "#/components/parameters/Range" }]
      responses:
        "200":
          description: Analytics report with top codes
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Analytics"
                  - type: object
                    properties:
                      top_qrcodes: { type: array, items: { $ref: "#/components/schemas/NamedCount" } }
  /campaigns:
    get:
      tags: [Campaigns]
      summary: List campaigns
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PerPage" }]
      responses:
        "200":
          description: A page of campaigns
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Paginated"
                  - type: object
                    properties:
                      data: { type: array, items: { $ref: "#/components/schemas/Campaign" } }
    post:
      tags: [Campaigns]
      summary: Create a campaign
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CampaignInput" }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /campaigns/{id}:
    parameters: [{ $ref: "#/components/parameters/Id" }]
    get:
      tags: [Campaigns]
      summary: Get a campaign
      responses:
        "200":
          description: The campaign
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
        "404": { $ref: "#/components/responses/NotFound" }
    patch:
      tags: [Campaigns]
      summary: Update a campaign
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CampaignInput" }
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Campaign" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
    delete:
      tags: [Campaigns]
      summary: Delete a campaign
      description: QR codes in the campaign are kept and unassigned.
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /campaigns/{id}/analytics:
    get:
      tags: [Campaigns, Analytics]
      summary: Campaign analytics
      parameters: [{ $ref: "#/components/parameters/Id" }, { $ref: "#/components/parameters/Range" }]
      responses:
        "200":
          description: Analytics report with per-code scans
          content:
            application/json:
              schema:
                allOf:
                  - $ref: "#/components/schemas/Analytics"
                  - type: object
                    properties:
                      per_qrcode: { type: array, items: { $ref: "#/components/schemas/NamedCount" } }
        "404": { $ref: "#/components/responses/NotFound" }
  /domains:
    get:
      tags: [Domains]
      summary: List custom domains
      responses:
        "200":
          description: Domains
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/Domain" } }
    post:
      tags: [Domains]
      summary: Add a custom domain
      description: Returns the DNS TXT record to create before calling verify. Session only.
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [hostname]
              properties:
                hostname: { type: string, examples: [qr.example.com] }
      responses:
        "201":
          description: Created (unverified)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Domain" }
        "409": { $ref: "#/components/responses/Conflict" }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /domains/{id}:
    delete:
      tags: [Domains]
      summary: Remove a custom domain
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }, { $ref: "#/components/parameters/CSRF" }]
      responses:
        "204": { description: Deleted }
        "404": { $ref: "#/components/responses/NotFound" }
  /domains/check:
    get:
      tags: [Domains]
      summary: On-demand TLS check for reverse proxies
      description: >-
        Returns 200 when a TLS certificate may be issued for `domain` (the primary host
        or a verified custom domain), 404 otherwise. Intended as the `ask` URL of
        Caddy's `on_demand_tls`. Public; returns no body.
      security: []
      parameters:
        - name: domain
          in: query
          required: true
          schema: { type: string }
      responses:
        "200": { description: Certificate may be issued }
        "404": { description: Unknown or unverified hostname }
  /domains/{id}/verify:
    post:
      tags: [Domains]
      summary: Verify domain ownership via DNS TXT lookup
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }, { $ref: "#/components/parameters/CSRF" }]
      responses:
        "200":
          description: Verified domain
          content:
            application/json:
              schema: { $ref: "#/components/schemas/Domain" }
        "404": { $ref: "#/components/responses/NotFound" }
        "422":
          description: TXT record not found (`verification_failed`)
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ErrorResponse" }
  /api-keys:
    get:
      tags: [API keys]
      summary: List API keys
      security: [{ session: [] }]
      responses:
        "200":
          description: Keys (never including the secret)
          content:
            application/json:
              schema:
                type: object
                properties:
                  data: { type: array, items: { $ref: "#/components/schemas/APIKey" } }
    post:
      tags: [API keys]
      summary: Create an API key
      description: The plaintext `key` is returned only in this response.
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/CSRF" }]
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [name]
              properties:
                name: { type: string, maxLength: 100 }
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                type: object
                properties:
                  api_key: { $ref: "#/components/schemas/APIKey" }
                  key: { type: string, examples: [qrf_4f3k9Zx0aLq2mN8pR1sT6uV7wY5bC3dE] }
        "422": { $ref: "#/components/responses/ValidationFailed" }
  /api-keys/{id}:
    delete:
      tags: [API keys]
      summary: Revoke an API key
      security: [{ session: [] }]
      parameters: [{ $ref: "#/components/parameters/Id" }, { $ref: "#/components/parameters/CSRF" }]
      responses:
        "204": { description: Revoked }
        "404": { $ref: "#/components/responses/NotFound" }
  /audit-log:
    get:
      tags: [Audit]
      summary: Your audit events (newest first)
      parameters: [{ $ref: "#/components/parameters/Page" }, { $ref: "#/components/parameters/PerPage" }]
      responses:
        "200":
          description: A page of events
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AuditPage" }
  /abuse-reports:
    post:
      tags: [Abuse]
      summary: Report an abusive QR code
      description: Public. `slug` may be a slug or a full redirect URL. Rate limited per IP.
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [slug, reason]
              properties:
                slug: { type: string }
                reason: { type: string, enum: [phishing, malware, spam, illegal, other] }
                details: { type: string, maxLength: 2000 }
                reporter_email: { type: [string, "null"], format: email }
      responses:
        "202":
          description: Received
          content:
            application/json:
              schema:
                type: object
                properties:
                  status: { type: string, const: received }
        "422": { $ref: "#/components/responses/ValidationFailed" }
        "429": { $ref: "#/components/responses/RateLimited" }
  /admin/overview:
    get:
      tags: [Admin]
      summary: Instance overview
      security: [{ session: [] }]
      responses:
        "200":
          description: Counts, storage, system health and traffic
          content:
            application/json:
              schema: { $ref: "#/components/schemas/AdminOverview" }
        "403": { $ref: "#/components/responses/Forbidden" }
...
components:
  securitySchemes:
    apiKey:
      type: http
      scheme: bearer
      bearerFormat: "qrf_<32 chars>"
      description: API key created in the dashboard.
    session:
      type: apiKey
      in: cookie
      name: qrforge_session
      description: Browser session. Unsafe requests also require the `X-CSRF-Token` header.
...